Where do I even start...
Rebasing the entire project to use KiCad 5.1.8 stable instead of nightly.
Most of the progress made on the schematic still in place, with a few exceptions that I can handle over the next few days.
Additonally. Resolves #17, #19, and parts of #18 (I still want to build some automated checks into the project once we are at a more advanced state)