Many, many changes...
Added a bunch of new footprints and a couple of new symbols.
Added the other side of the CM4 connector and wired its features.
Added some more detailed models for a fair few components.
Added a new schematic to handle power for USB ports.
Updated the PCB with the new components and moved some around to populate the rear IO panel.